Rename `general` components dir to `ui`
