+++ src/components/layout/search/search-button.tsx
@@ -1,7 +1,7 @@
 import type { ComponentPropsWithoutRef } from 'react';
 
 import { SearchBoxButton } from './search-box-button';
-import { IconButton } from '@/components/general';
+import { IconButton } from '@/components/ui';
 import { SearchIcon } from '@/assets/icons';
 
 export const SearchButton = ({
